+++ src/gas.ts
@@ -1,0 +1,31 @@
+export enum Gas {
+    'Fullerite-C28' = 30375,
+    'Fullerite-C32' = 30376,
+    'Fullerite-C320' = 30377,
+    'Fullerite-C50' = 30370,
+    'Fullerite-C540' = 30378,
+    'Fullerite-C60' = 30371,
+    'Fullerite-C70' = 30372,
+    'Fullerite-C72' = 30373,
+    'Fullerite-C84' = 30374,
+
+    'Amber Cytoserocin' = 25268,
+    'Amber Mykoserocin' = 28694,
+    'Azure Cytoserocin' = 25279,
+    'Azure Mykoserocin' = 28695,
+    'Celadon Cytoserocin' = 25275,
+    'Celadon Mykoserocin' = 28696,
+    'Chartreuse Cytoserocin' = 28630, // Unavailable on market.
+    'Gamboge Cytoserocin' = 28629, // Unavailable on market.
+    'Golden Cytoserocin' = 25273,
+    'Golden Mykoserocin' = 28697,
+    'Hiemal Tricarboxyl Vapor' = 49787, // In Materials - Raw Materials - Standard Ores - Veldspar ?
+    'Lime Cytoserocin' = 25277,
+    'Lime Mykoserocin' = 28698,
+    'Malachite Cytoserocin' = 25276,
+    'Malachite Mykoserocin' = 28699,
+    'Vermillion Cytoserocin' = 25278,
+    'Vermillion Mykoserocin' = 28700,
+    'Viridian Cytoserocin' = 25274,
+    'Viridian Mykoserocin' = 28701,
+}