+++ src/mineral.ts
@@ -1,0 +1,10 @@
+export enum Mineral {
+    Tritanium = 34,
+    Pyerite = 35,
+    Mexallon = 36,
+    Isogen = 37,
+    Nocxium = 38,
+    Zydrine = 39,
+    Megacyte = 40,
+    Morphite = 11399,
+}