+++ src/ore.ts
@@ -1,0 +1,80 @@
+export enum Ore {
+    'Veldspar' = 1230,
+    'Concentrated Veldspar' = 17470,
+    'Dense Veldspar' = 17471,
+    'Stable Veldspar' = 46689,
+
+    'Scordite' = 1228,
+    'Condensed Scordite' = 17463,
+    'Massive Scordite' = 17464,
+    'Glossy Scordite' = 46687,
+
+    'Pyroxeres' = 1224,
+    'Solid Pyroxeres' = 17459,
+    'Viscous Pyroxeres' = 17460,
+    'Opulent Pyroxeres' = 46686,
+
+    'Plagioclase' = 18,
+    'Azure Plagioclase' = 17455,
+    'Rich Plagioclase' = 17456,
+    'Sparkling Plagioclase' = 46685,
+
+    'Omber' = 1227,
+    'Silvery Omber' = 17867,
+    'Golden Omber' = 17868,
+    'Platinoid Omber' = 46684,
+
+    'Kernite' = 20,
+    'Luminous Kernite' = 17452,
+    'Fiery Kernite' = 17453,
+    'Resplendant Kernite' = 46683,
+
+    'Jaspet' = 1226,
+    'Pure Jaspet' = 17448,
+    'Pristine Jaspet' = 17449,
+    'Immaculate Jaspet' = 46682,
+
+    'Hemorphite' = 1231,
+    'Vivid Hemorphite' = 17444,
+    'Radiant Hemorphite' = 17445,
+    'Scintillating Hemorphite' = 46681,
+
+    'Hedbergite' = 21,
+    'Vitric Hedbergite' = 17440,
+    'Glazed Hedbergite' = 17441,
+    'Lustrous Hedbergite' = 46680,
+
+    'Gneiss' = 1229,
+    'Iridescent Gneiss' = 17865,
+    'Prismatic Gneiss' = 17866,
+    'Brilliant Gneiss' = 46679,
+
+    'Dark Ochre' = 1232,
+    'Onyx Ochre' = 17436,
+    'Obsidian Ochre' = 17437,
+    'Jet Ochre' = 46675,
+
+    'Spodumain' = 19,
+    'Bright Spodumain' = 17466,
+    'Gleaming Spodumain' = 17467,
+    'Dazzling Spodumain' = 46688,
+
+    'Crokite' = 1225,
+    'Sharp Crokite' = 17432,
+    'Crystalline Crokite' = 17433,
+    'Pellucid Crokite' = 46677,
+
+    'Bistot' = 1223,
+    'Triclinic Bistot' = 17428,
+    'Monoclinic Bistot' = 17429,
+    'Cubic Bistot' = 46676,
+
+    'Arkonor' = 22,
+    'Crimson Arkonor' = 17425,
+    'Prime Arkonor' = 17426,
+    'Flawless Arkonor' = 46678,
+
+    'Mercoxit' = 11396,
+    'Magma Mercoxit' = 17869,
+    'Vitreous Mercoxit' = 17870,
+}